faifly/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyascii==0.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.5
django-filter==21.1
django-modelcluster==6.0
django-permissionedforms==0.1
django-taggit==2.1.0
django-treebeard==4.5.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
frozenlist==1.3.0
html5lib==1.1
idna==3.3
l18n==2021.3
loader==2017.9.11
Logbook==1.5.3
multidict==6.0.2
openpyxl==3.0.10
pendulum==2.1.2
Pillow==9.1.1
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.1
pytzdata==2020.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tablib==3.2.1
telepath==0.2
termcc==2019.8.2
terminaltables==3.1.10
urllib3==1.26.9
wagtail==3.0.1
webencodings==0.5.1
Willow==1.4.1
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
yarl==1.7.2
